withers/pyproject.toml
[project]
name = "withers"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.1.5",
    "djangorestframework>=3.15.2",
    "drf-spectacular>=0.28.0",
]
